requirements.txt
greenlet>=1.1.2
iteround>=1.0.3
more-itertools>=8.12.0
numpy>=1.22.3
pandas>=1.4.1
psycopg2-binary>=2.9.3
python-dateutil>=2.8.2
pytz>=2021.3
six>=1.16.0
tqdm>=4.63.1
openpyxl>=3.0.9
xlrd
click
loguru